requirements.txt
customtkinter==5.2.1
darkdetect==0.8.0
packaging==23.2
Pillow==10.1.0
tk==0.1.0
yt-dlp==2025.1.26